tests wip

+++ AmericanDream/Model/ChangeRateCacheManager.swift
@@ -8,11 +8,12 @@
 import Foundation
 
 class ChangeRateCacheManager {
-    func load(key:String) -> [String:Float] {
-        return UserDefaults.standard.object(forKey: key) as? [String:Float] ?? [:]
+
+    func load(key:String) -> [String:Double] {
+        return UserDefaults.standard.object(forKey: key) as? [String:Double] ?? [:]
     }
 
-    func save(key:String, value:[String:Float]) {
+    func save(key:String, value:[String:Double]) {
         UserDefaults.standard.set(value, forKey: key)
     }
 }
